requirements.txt
streamlit
numpy
opencv-python
asyncio
boto3
Pillow
python-dotenv
boto3
